Powershell script to set userhome folder permissions
The script configures acl permissions for user home folders. It uses the
concept of a template folder from which all the permissions will be
copied, and it also adds the user to the ace and confgures the acl for
that user. Example, I have a user called John Doe, his home folder is
called jdoe. The script reads the foldername and then converts it to the
jdoe@domainname.com format so it can create the ace and acl.